packages/core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prlens-core"
version = "0.1.8"
description = "Core review engine for prlens — AI-powered GitHub PR code reviewer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "PyGithub>=2.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.25"]
openai = ["openai>=1.0"]
all = ["anthropic>=0.25", "openai>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "black>=24.0",
    "flake8>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
prlens_core = ["guidelines/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
